requirements.txt
black
flake8
glfw
isort
numpy
PyOpenGL
PyOpenGL-accelerate
scipy
